Update link_builder.html.erb
We need to have a way to know what type of donation it is.
Tithe, Offering, Building Fund, Youth Ministry, Kitchen Renovation's, Missions, Cambodian Orphanage, Other:______
I adjusted it so that instead of "Order Number" - It reads 'Designation (Unspecified Amounts Will Be Considered Tithe)
Then underneath I included details of the types of Designations
( Tithe, Offering, Building Fund, Youth Ministry, Kitchen Renovation's, Missions, Cambodian Orphanage, Other:______ )
The form can have the option for people to enter in the amount of their giving and then manually enter in the type, as an 'order number' so to speak.
However - if there was a way for multiple types of giving to be tallied together with one transaction, with details as to the amount per type of donation (exactly like an offering envelope).  
If someone there could create a form that has the titles of the types of transactions listed above, with the ability to enter dollar amounts in each of them, and have a grand total for 1 transaction, while keeping the information regarding each designation - that would be great. 
We only need the Name, Email, Address and Designations on the form
Thank you